requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
flask-wtf==1.2.1
werkzeug==3.0.1
flask-login==0.6.3